+++ CommHandler/src/test/java/com/smartgridready/communicator/modbus/impl/IntrospectiveDeviceTester.java
@@ -5,8 +5,8 @@ import com.smartgridready.ns.v0.DeviceFrame;
 import com.smartgridready.communicator.common.helper.DeviceDescriptionLoader;
 import com.smartgridready.communicator.common.helper.DriverFactoryLoader;
 import com.smartgridready.communicator.common.impl.SGrDeviceBase;
-import com.smartgridready.communicator.rest.http.client.ApacheHttpRequestFactory;
 import com.smartgridready.communicator.rest.impl.SGrRestApiDevice;
+import com.smartgridready.driver.api.http.GenHttpClientFactory;
 import com.smartgridready.driver.api.modbus.GenDriverAPI4Modbus;
 import com.smartgridready.driver.api.modbus.GenDriverAPI4ModbusFactory;
 
@@ -142,7 +142,8 @@ public class IntrospectiveDeviceTester {
                 "SGr_02_mmmm_8288089799_Smart-me_SubMeterElectricity_V1.0.0.xml",
                 properties);
 
-        SGrRestApiDevice restApiDevice = new SGrRestApiDevice(devDesc, new ApacheHttpRequestFactory());
+        GenHttpClientFactory factory = DriverFactoryLoader.getImplementation(GenHttpClientFactory.class);
+        SGrRestApiDevice restApiDevice = new SGrRestApiDevice(devDesc, factory);
         return new Tuple3<>(devDesc, restApiDevice, null);
     }
 
@@ -159,7 +160,8 @@ public class IntrospectiveDeviceTester {
                 "SGr_02_0018_CLEMAP_EIcloudEnergyMonitor_V1.0.0.xml",
                 properties);
 
-        SGrRestApiDevice restApiDevice = new SGrRestApiDevice(devDesc, new ApacheHttpRequestFactory());
+        GenHttpClientFactory factory = DriverFactoryLoader.getImplementation(GenHttpClientFactory.class);
+        SGrRestApiDevice restApiDevice = new SGrRestApiDevice(devDesc, factory);
         return new Tuple3<>(devDesc, restApiDevice, null);
     }
 
@@ -171,7 +173,8 @@ public class IntrospectiveDeviceTester {
         DeviceDescriptionLoader loader = new DeviceDescriptionLoader();
         DeviceFrame devDesc = loader.load("", deviceDescUrl != null ? deviceDescUrl.getPath() : null);
 
-        SGrRestApiDevice restApiDevice = new SGrRestApiDevice(devDesc, new ApacheHttpRequestFactory());
+        GenHttpClientFactory factory = DriverFactoryLoader.getImplementation(GenHttpClientFactory.class);
+        SGrRestApiDevice restApiDevice = new SGrRestApiDevice(devDesc, factory);
 
         Properties parameters = new Properties();
         parameters.put("point", "CH1018601234500000000000000011642");
@@ -189,7 +192,8 @@ public class IntrospectiveDeviceTester {
         DeviceDescriptionLoader loader = new DeviceDescriptionLoader();
         DeviceFrame devDesc = loader.load("", deviceDescUrl != null ? deviceDescUrl.getPath() : null);
 
-        SGrRestApiDevice restApiDevice = new SGrRestApiDevice(devDesc, new ApacheHttpRequestFactory());
+        GenHttpClientFactory factory = DriverFactoryLoader.getImplementation(GenHttpClientFactory.class);
+        SGrRestApiDevice restApiDevice = new SGrRestApiDevice(devDesc, factory);
 
         Properties parameters = new Properties();
         parameters.put("start_timestamp", "2023-09-06T00:00:00+02:00");